+++ packages/browser-relay/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- Tabs opened through the relay now open in the background without stealing focus from the user's foreground tab ([#11662](https://github.com/can1357/oh-my-pi/pull/11662) by [@jwaldrip](https://github.com/jwaldrip)).
+- Activating a driven tab for screenshots no longer raises or focuses the browser window ([#11662](https://github.com/can1357/oh-my-pi/pull/11662) by [@jwaldrip](https://github.com/jwaldrip)).
+
 ## [18.0.7] - 2026-08-26
 
 ### Changed